Move controls to separate files where possible
