+++ iui/src/controls/mod.rs
@@ -17,8 +17,16 @@ mod window;
 pub use self::window::*;
 mod layout;
 pub use self::layout::*;
-mod entry;
-pub use self::entry::*;
+mod textentry;
+pub use self::textentry::*;
+mod numericentry;
+pub use self::numericentry::*;
+mod checkbox;
+pub use self::checkbox::*;
+mod combobox;
+pub use self::combobox::*;
+mod radiobuttons;
+pub use self::radiobuttons::*;
 mod progressbar;
 pub use self::progressbar::*;
 mod area;
